requirements.txt
Django  # ==1.7
django-debug-toolbar  # ==1.4
django-localflavor  # ==1.2
selenium  # ==2.53.1
sqlparse  # ==0.1.19
django-environ  # ==0.4.0
